move src files to project folder
