+++ manm_cs/noise/__init__.py
