+++ manm_cs/noise/continous_noise.py
@@ -1,5 +1,5 @@
-from src.manm_cs.noise.noise import Noise
-from src.manm_cs.prob_distributions.continuous import ContinuousDistribution
+from manm_cs.noise.noise import Noise
+from manm_cs.prob_distributions.continuous import ContinuousDistribution
 
 
 class ContinuousNoise(Noise):
