+++ manm_cs/prob_distributions/discrete/__init__.py
