+++ manm_cs/prob_distributions/discrete/uniform_discrete_distribution.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from src.manm_cs.prob_distributions.discrete.discrete_distribution import DiscreteDistribution
+from manm_cs.prob_distributions.discrete.discrete_distribution import DiscreteDistribution
 
 
 class UniformDiscreteDistribution(DiscreteDistribution):
