+++ manm_cs/prob_distributions/prob_distribution.py
@@ -2,7 +2,7 @@ from abc import ABC
 
 import numpy as np
 
-from src.manm_cs.variables import VariableType
+from manm_cs.variables import VariableType
 
 
 class ProbDistribution(ABC):
