+++ manm_cs/variables/variable_type.py
