--- src/manm_cs/prob_distributions/__init__.py
@@ -1,4 +1,0 @@
-from src.manm_cs.prob_distributions.continuous import GaussianDistribution
-from src.manm_cs.prob_distributions.discrete import BinomialDistribution
-from src.manm_cs.prob_distributions.discrete import CustomDiscreteDistribution
-from src.manm_cs.prob_distributions.discrete import UniformDiscreteDistribution
